requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
pillow>=8.0.0
pywin32>=300
pytesseract>=0.3.8